Creating WaterObjects
Made code changes in MainActivity and HomeFragments classes. App will now create new WaterObject when user has defined user's weight. I noticed because of these changes the preferenceFragment class didn't work so made some changes in there also in some target functions. Now the app doesn't crash but it show wrong weight which is a bit odd. The app doesn't save the object in memory so we need to fix that..